==> Final/Species_distribution_modelling/02.batch_env_swd_prep.r <==
#### Batch Script to prepare future environmental data for modelling
#### C. James, L.Hodgson...............9th January 2012

futdir = "/home/jc165798/Climate/CIAS/Australia/5km/monthly_csv/"

ESs = list.files(futdir, pattern="RCP") # list the emission scenarios
GCMs = list.files(paste(futdir,ESs[1],sep='')) # get a list of GCMs

sh.dir='/home/jc246980/SDM/temp/'; setwd(sh.dir)
for (es in ESs) {
for (gcm in GCMs) {
#for trial runs
#es='RCP85'
#gcm='cccma-cgcm31'
#yy='2085'

zz = file(paste(es,'.',gcm,'.prep_data.sh',sep=''),'w')
cat('#!/bin/bash\n',file=zz)
cat('cd $PBS_O_WORKDIR\n',file=zz)
cat('source /etc/profile.d/modules.sh\n',file=zz)
cat('module load R/2.15.1\n',file=zz)
cat("R CMD BATCH --no-save --no-restore '--args es=\"",es,"\" gcm=\"",gcm,"\" ' /home/jc246980/Freshwater_scripts/Final/Species_distribution_modelling/02.run_env_swd_prep.r ",es,'.',gcm,'.Rout \n',sep='',file=zz) #run the R script in the background
close(zz)

##submit the script
system(paste('qsub -m n -l nodes=1:ppn=5 ',es,'.',gcm,'.prep_data.sh',sep=''))
}
}
